+++ tests/testthat/test-read_transect.R
@@ -5,3 +5,12 @@ test_that("Checking read_transect class", {
   expect_is(result, 'data.frame')
 
 })
+
+test_that("Checking read_transect class", {
+
+  result <- read_transect(training = TRUE)
+
+  expect_is(result, 'data.frame')
+
+})
+
