### 2021-05-04 #3243 @fawda123
@kthyng @leouieda here's a list of potential reviewers by username.

From the [list of potential reviewers](https://bit.ly/joss-reviewers): jhollist, TomasBeuzen, paleolimbot

Not from the list of potential reviewers, but other good options: sheilasaia, jsta, padilla410, mishafredmeyer
