+++ paper/paper.md
@@ -48,7 +48,7 @@ journal: JOSS
 
 Many environmental programs report on the status and trends of natural resources to inform management decisions for protecting or restoring environmental condition.  The National Estuary Program (NEP) in the United States is one example of a resource management institution focused on "estuaries of national significance" that provides place-based solutions to managing coastal resources.  There are 28 NEPs in the United States, each with similar but location-specific programmatic goals to address environmental challenges related to water quality, alteration of hydrologic flows, invasive species, climate change,  declines in fish and wildlife populations, pathogens and other contaminants, and stormwater management.  A critical need of each NEP is the synthesis of data from disparate sources that can inform management response to address these environmental challenges. 
 
-The Tampa Bay Estuary Program (TBEP) in Florida, USA is responsible for developing and implementing a place-based plan to sustain historical and future progress in the restoration of Tampa Bay [@tbep1017].  The needs of TBEP for reporting on indicators of environmental condition are similar to other environmental organizations.  Multiple local and regional partners collect data that are used for different reporting products.  Without data synthesis tools that are transparent, accessible, and reproducible, NEP staff and colleagues waste time and resources compiling information by hand.  The `tbeptools` R software package can be used for routine development of reporting products, allowing for more efficient use of limited resources and a more effective approach to communicate research to environmental decision-makers.     
+The Tampa Bay Estuary Program (TBEP) in Florida, USA is responsible for developing and implementing a place-based plan to sustain historical and future progress in the restoration of Tampa Bay [@tbep1017].  The needs of TBEP for reporting on indicators of environmental condition are similar to other environmental organizations.  Multiple local and regional partners collect data that are used for different reporting products.  Without data synthesis tools that are transparent, accessible, and reproducible, NEP staff and colleagues waste time and resources compiling information by hand.  The `tbeptools` R software package can be used for routine development of reporting products, allowing for more efficient use of limited resources and a more effective approach to communicate research to environmental decision-makers.  The package also imports required data directly from sources, removing the need to manually obtain information prior to reporting.    
 
 # Statement of need
 
