### 2021-05-05 9170014
attempting functions by concept again, reorg intro vignette with content to readme
