+++ R/read_dlcurrent.R
@@ -10,7 +10,7 @@
 #'
 #' @export
 #'
-#' @family read
+#' @concept read
 #'
 #' @details The local copy is checked against a temporary file downloaded from the location specified by \code{urlin}.  The local file is replaced with the downloaded file if the MD5 hashes are different.
 #' @examples
