+++ R/show_transectsum.R
@@ -13,7 +13,7 @@
 #'
 #' @importFrom dplyr %>%
 #'
-#' @family visualize
+#' @concept show
 #'
 #' @examples
 #' \dontrun{
