### 2021-05-13 #3243 @kthyng
Hi @paleolimbot and @sheilasaia! Are you interested in reviewing this submission for the Journal for Open Source Software? This is an open review which takes place on github with issues. You can learn more about the process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. I would hope you could start the review in the next few weeks with the intention of being through the typically iterative process in 4-6 weeks. Thanks for your consideration!
