### 2021-05-18 #3243 @kthyng
Hi @TomasBeuzen! Are you interested in reviewing this submission for JOSS? This is an open review which takes place on github with issues. You can learn more about the process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. I would hope you could start the review in the next few weeks with the intention of being through the typically iterative process in 4-6 weeks. Thanks for your consideration! Also I LOVE the readme on your github page! How cool! I've marked it to emulate when I can get to it.
