### 2021-05-20 #3243 @TomasBeuzen
Hi @kthyng - thanks for the offer (and the kudos on my readme!). I'll decline this time for two reasons: 1. I've just moved countries and am trying to get my life in order (so trying to say "no" to these extra things), 2. For the last 2 years I've been teaching Python so am not sure my R knowledge is up to scratch for this review! Thanks again for the offer and best of luck with the review process!
