### 2021-05-27 460ae0a
New DO calculation function for anlz_tdlcrkindic
