### 2021-06-01 58e4fa6
added dosat exceedances to anlz_tdlcrkindic
