+++ R/anlz_tdlcrkindic.R
@@ -6,7 +6,11 @@
 #' @param iwrraw FDEP impaired waters rule run 56 data base as \code{\link{data.frame}}
 #' @param yr numeric for reference year to evaluate, scores are based on the planning period beginning ten years prior to this date
 #'
-#' @details Annual geometric means for additional water quality data available at each wbid, JEI combination.  Florida trophic state index values are also estimated where data are available.  Nitrogen ratios are estimated for JEIs that cover source (upstream, freshwater) and tidal (downstream) WBIDs, defined as the ratio of concentrations between the two (i.e., ratios > 1 mean source has higher concentrations).  Nitrogen ratios for a given year reflect the ratio of the median nitrogen concentrations when they were measured in both a source and tidal segment during the same day.  Note that a ratio of one can be obtained if both the source and tidal segments are at minimum detection.
+#' @details Annual geometric means for additional water quality data available at each wbid, JEI combination.  Florida trophic state index values are also estimated where data are available.
+#'
+#' Nitrogen ratios are estimated for JEIs that cover source (upstream, freshwater) and tidal (downstream) WBIDs, defined as the ratio of concentrations between the two (i.e., ratios > 1 mean source has higher concentrations).  Nitrogen ratios for a given year reflect the ratio of the median nitrogen concentrations when they were measured in both a source and tidal segment during the same day.  Note that a ratio of one can be obtained if both the source and tidal segments are at minimum detection.
+#'
+#' Indicators for years where more than 10\% of observations exceed DO saturation criteria are also estimated.  The \code{do_bnml} and \code{do_prop} columns show a 1 or 0 for a given year to indicate if more than ten percent of observations were below DO percent saturation of 42.  The first column is based on a binomial probability exceedance that takes into account the number of observations in the year and the second column is based on a simple proportional estimate from the raw data.
 #'
 #' @return A \code{\link{data.frame}} with the indicator values for each tidal creek
 #' @export
@@ -91,9 +95,48 @@ anlz_tdlcrkindic <- function(tidalcreeks, iwrraw, yr = 2018) {
       no23_ratio = no23_source / no23_tidal
     )
 
+  # estimate DO exceedance
+  # returns a column do_bnml as dep method based on likelihood of obs dosat < 42 in a given year
+  # returns a column do_prop as straight calc of proportion of days with dosat < 42 in a given year
+  do_exceed <- iwrraw %>%
+    dplyr::filter(masterCode %in% c('DO','DOSAT','TEMP','SALIN')) %>%
+    dplyr::group_by(class, JEI, wbid, sta, year, month, day, masterCode) %>%
+    dplyr::summarise(
+      result = mean(result, na.rm = T),
+      .groups = 'drop'
+      ) %>%
+    dplyr::ungroup() %>%
+    tidyr::spread(masterCode, result) %>%
+    dplyr::filter(!is.na(DOSAT))%>%
+    dplyr::mutate(
+      do_cnt = dplyr::case_when(
+        DOSAT < 42 ~ 1,
+        DOSAT >= 42 ~ 0
+        )
+      ) %>%
+    dplyr::group_by(JEI,wbid,year) %>%
+    dplyr::summarise(
+      cnt = n(),
+      DOSAT_prop = mean(do_cnt, na.rm = T),
+      .groups = 'drop'
+      ) %>%
+    dplyr::mutate(
+      xcd = DOSAT_prop * cnt,
+      bnml = 1 - pbinom(xcd - 1, cnt, 0.1),
[…]
