+++ man/seagrass.Rd
@@ -22,10 +22,10 @@ Seagrass coverage by year
 \dontrun{
 
 seagrass <- structure(list(
-   Year = c(1950L, 1982L, 1988L, 1990L, 1992L, 1994L, 1996L, 1999L, 2001L, 2004L, 2006L, 2008L, 2010L, 2012L, 2014L, 2016L, 2018L),
-   Acres = c(40420, 21650, 23285, 25226, 25753, 26518, 26916, 24841, 26078, 27021, 28299, 29647, 32897, 34642, 40294.71, 41655.16, 40651.55),
-   Hectares = c(16357.39, 8761.44, 9423.11, 10208.6, 10421.87, 10731.45, 10892.52, 10052.8, 10553.39, 10935.01, 11452.2, 11997.72, 13312.94, 14019.27, 16306.69, 16857.25, 16451.1
- )), class = "data.frame", row.names = c(NA, -17L))
+   Year = c(1950L, 1982L, 1988L, 1990L, 1992L, 1994L, 1996L, 1999L, 2001L, 2004L, 2006L, 2008L, 2010L, 2012L, 2014L, 2016L, 2018L, 2020L),
+   Acres = c(40420, 21650, 23285, 25226, 25753, 26518, 26916, 24841, 26078, 27021, 28299, 29647, 32897, 34642, 40294.71, 41655.16, 40651.55, 33944.94),
+   Hectares = c(16357.39, 8761.44, 9423.11, 10208.6, 10421.87, 10731.45, 10892.52, 10052.8, 10553.39, 10935.01, 11452.2, 11997.72, 13312.94, 14019.27, 16306.69, 16857.25, 16451.1, 13737.08)
+ ), class = "data.frame", row.names = c(NA, -18L))
 
 save(seagrass, file = 'data/seagrass.RData', compress = 'xz')
 
