### 2021-06-07 bc82756
change hex file name to logo
