### 2021-06-07 9aa2fa2
removed mapview 2.7.8 depends for gh actions
