+++ .github/workflows/pkgdown.yaml
@@ -40,7 +40,6 @@ jobs:
       - name: Install dependencies
         run: |
           remotes::install_deps(dependencies = TRUE)
-          remotes::install_version("mapview", version = "2.7.8")
           remotes::install_github("fawda123/Jabbrev")
           install.packages("pkgdown", type = "binary")
         shell: Rscript {0}
