+++ .github/workflows/R-CMD-check.yaml
@@ -73,7 +73,6 @@ jobs:
       - name: Install dependencies
         run: |
           remotes::install_deps(dependencies = TRUE)
-          remotes::install_version("mapview", version = "2.7.8")
           remotes::install_github("fawda123/Jabbrev")
           remotes::install_cran("rcmdcheck")
         shell: Rscript {0}
