+++ .github/workflows/pkgdown.yaml
@@ -8,7 +8,7 @@ name: pkgdown
 
 jobs:
   pkgdown:
-    runs-on: macOS-latest
+    runs-on: ubuntu-20.04
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
     steps:
@@ -32,10 +32,18 @@ jobs:
           key: ${{ runner.os }}-${{ hashFiles('.github/R-version') }}-1-${{ hashFiles('.github/depends.Rds') }}
           restore-keys: ${{ runner.os }}-${{ hashFiles('.github/R-version') }}-1-
 
-      - name: Install OS depends
+      - name: Install system dependencies
+        if: runner.os == 'Linux'
         run: |
-          brew install gdal
-          brew install udunits
+          while read -r cmd
+          do
+            eval sudo $cmd
+          done < <(Rscript -e 'writeLines(remotes::system_requirements("ubuntu", "20.04"))')
+
+      - name: Install Ubuntu depends
+        run: |
+          sudo apt-get install libgdal-dev
+          sudo apt-get install udunits-bin
 
       - name: Install dependencies
         run: |
