+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: tbeptools
 Title: Data and Indicators for the Tampa Bay Estuary Program
-Version: 1.0.0
-Date: 2021-06-01
+Version: 1.0.1
+Date: 2021-06-11
 Authors@R: c(
     person(given = "Marcus",
       family = "Beck",
