### 2021-06-14 61ff412
Added summarization of annual indicator scores for radar charts as dataframe out2. For some reason spread function wasnt working until i declared tidyr on line 83 but probably something on my end.
