### 2021-06-14 #3243 @ibarraespinosa
Hello @kthyng and thanks for the invitation. Unfortunately, I will only have time in August.
