### 2021-06-14 df74a2a
i think this handles the na's better in assigning the boolean (1/0) to the indicators. Also, i think i may have added a dependency (stringr) by using the str_detect function on line 147?
