### 2021-06-18 9ecff1b
minor fix to anlz_tdlcrkindic for radar plot output, doc update
