+++ R/show_tdlcrkradar.R
@@ -1,0 +1,233 @@
+#' Radar plots for tidal creek indicators
+#'
+#' Radar plots for tidal creek indicators
+#'
+#' @param id numeric indicating the \code{id} number of the tidal creek to plot
+#' @param cntdat output from \code{\link{anlz_tdlcrkindic}}
+#'
+#' @return A radar plot
+#' @export
+#'
+#' @details See details in \code{\link{anlz_tldcrkindic}} for an explanation of the indicators
+#'
+#' Internal code borrowed heavily from \code{\link[fmsb]{radarchart}}.
+#'
+#' @concept show
+#'
+#' @examples
+#' cntdat <- anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2018, radar = TRUE)
+#' show_tdlcrkradar(35, cntdat)
+show_tdlcrkradar <- function (id, cntdat, col = '#338080E6'){
+
+  axistype <- 3
+  seg <- 5
+  pty <- 16
+  plty <- 1
+  plwd <- 5
+  pdensity <- NULL
+  pangle <- 45
+  pfcol <- ggplot2::alpha(col, 0.4)
+  cglty <- 1
+  cglwd <- 1
+  cglcol <- "grey"
+  axislabcol <- "grey"
+  title <- ""
+  maxmin <- TRUE
+  na.itp <- TRUE
+  centerzero <- FALSE
+  vlabels <- NULL
+  vlcex <- 0.8
+  caxislabels <- seq(0,100,20)
+  calcex <- NULL
+  paxislabels <- ''
+  palcex <- NULL
+
+  # prep
+  toplo <- cntdat %>%
+    dplyr::filter(id == !!id) %>%
+    dplyr::select(JEI, var, val) %>%
+    dplyr::mutate(
+      var = factor(var,
+        levels = c('chla_ind', 'tsi_ind', 'tn_ind', 'nox_ind', 'do_prop', 'ch_tn_rat_ind'),
+        labels = c('Chloropyll', 'TSI score', 'Total Nitrogen', 'Nitrate Ratio', 'DO Saturation', 'Chl/TN Ratio'))
+    ) %>%
+    tidyr::spread(var, val) %>%
+    tibble::column_to_rownames(var = 'JEI') %>%
+    rbind(rep(0, nrow(.)), .) %>%
+    rbind(rep(100, nrow(.)), .)
+
+  if(nrow(toplo) == 0)
+    stop('No marine segments in id')
+
+  plot(c(-1.2, 1.2), c(-1.2, 1.2), type = "n", frame.plot = FALSE,
+       axes = FALSE, xlab = "", ylab = "", main = title, asp = 1)
+
+  theta <- seq(90, 450, length = n + 1) * pi/180
+  theta <- theta[1:n]
+  xx <- cos(theta)
+  yy <- sin(theta)
+  CGap <- ifelse(centerzero, 0, 1)
+  for (i in 0:seg) {
+    polygon(xx * (i + CGap)/(seg + CGap), yy * (i + CGap)/(seg +
+                                                             CGap), lty = cglty, lwd = cglwd, border = cglcol)
+    if (axistype == 1 | axistype == 3)
+      CAXISLABELS <- paste(i/seg * 100, "(%)")
+    if (axistype == 4 | axistype == 5)
+      CAXISLABELS <- sprintf("%3.2f", i/seg)
+    if (!is.null(caxislabels) & (i < length(caxislabels)))
+      CAXISLABELS <- caxislabels[i + 1]
+    if (axistype == 1 | axistype == 3 | axistype == 4 |
+        axistype == 5) {
+      if (is.null(calcex))
[…]
