### 2021-06-20 5649849
cleaning up show_tdlcrkradar, #41
