+++ R/show_tdlcrkradar.R
@@ -21,26 +21,10 @@ show_tdlcrkradar <- function (id, cntdat, col = '#338080E6'){
 
   axistype <- 3
   seg <- 5
-  pty <- 16
-  plty <- 1
-  plwd <- 5
-  pdensity <- NULL
-  pangle <- 45
   pfcol <- ggplot2::alpha(col, 0.4)
-  cglty <- 1
-  cglwd <- 1
   cglcol <- "grey"
   axislabcol <- "grey"
-  title <- ""
-  maxmin <- TRUE
-  na.itp <- TRUE
-  centerzero <- FALSE
-  vlabels <- NULL
-  vlcex <- 0.8
-  caxislabels <- seq(0,100,20)
-  calcex <- NULL
-  paxislabels <- ''
-  palcex <- NULL
+  caxislabels <- seq(0, 100, 20)
 
   # prep
   toplo <- cntdat %>%
@@ -60,174 +44,88 @@ show_tdlcrkradar <- function (id, cntdat, col = '#338080E6'){
     stop('No marine segments in id')
 
   plot(c(-1.2, 1.2), c(-1.2, 1.2), type = "n", frame.plot = FALSE,
-       axes = FALSE, xlab = "", ylab = "", main = title, asp = 1)
+       axes = FALSE, xlab = "", ylab = "", asp = 1)
 
+  n <- ncol(toplo)
   theta <- seq(90, 450, length = n + 1) * pi/180
   theta <- theta[1:n]
   xx <- cos(theta)
   yy <- sin(theta)
-  CGap <- ifelse(centerzero, 0, 1)
+  CGap <- 1
   for (i in 0:seg) {
-    polygon(xx * (i + CGap)/(seg + CGap), yy * (i + CGap)/(seg +
-                                                             CGap), lty = cglty, lwd = cglwd, border = cglcol)
-    if (axistype == 1 | axistype == 3)
-      CAXISLABELS <- paste(i/seg * 100, "(%)")
-    if (axistype == 4 | axistype == 5)
-      CAXISLABELS <- sprintf("%3.2f", i/seg)
-    if (!is.null(caxislabels) & (i < length(caxislabels)))
-      CAXISLABELS <- caxislabels[i + 1]
-    if (axistype == 1 | axistype == 3 | axistype == 4 |
-        axistype == 5) {
-      if (is.null(calcex))
-        text(-0.05, (i + CGap)/(seg + CGap), CAXISLABELS,
-             col = axislabcol)
-      else text(-0.05, (i + CGap)/(seg + CGap), CAXISLABELS,
-                col = axislabcol, cex = calcex)
-    }
+
+    polygon(xx * (i + CGap)/(seg + CGap), yy * (i + CGap)/(seg + CGap), lty = 1, lwd = 1, border = cglcol)
+
+    CAXISLABELS <- caxislabels[i + 1]
+
+    text(-0.05, (i + CGap)/(seg + CGap), CAXISLABELS,
+         col = axislabcol)
+
   }
-  if (centerzero) {
-    arrows(0, 0, xx * 1, yy * 1, lwd = cglwd, lty = cglty,
-           length = 0, col = cglcol)
-  }
-  else {
-    arrows(xx/(seg + CGap), yy/(seg + CGap), xx * 1, yy *
-             1, lwd = cglwd, lty = cglty, length = 0, col = cglcol)
-  }
-  PAXISLABELS <- toplo[1, 1:n]
-  if (!is.null(paxislabels))
-    PAXISLABELS <- paxislabels
-  if (axistype == 2 | axistype == 3 | axistype == 5) {
-    if (is.null(palcex))
[…]
