+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: tbeptools
 Title: Data and Indicators for the Tampa Bay Estuary Program
 Version: 1.0.2
-Date: 2021-06-18
+Date: 2021-06-20
 Authors@R: c(
     person(given = "Marcus",
       family = "Beck",
