### 2021-06-21 66ba1d6
+updated tidalcreekreport graphic and vignette to better reflect manuscript. citation for manuscript provided.
