+++ tests/testthat/test-anlz_hydroload.R
@@ -1,3 +1,8 @@
-test_that("Checking error on anlz_hydroload", {
-  expect_error(anlz_hydroload(2019, 'asdf'))
+test_that("Checking class on anlz_hydroload", {
+
+  noaa_key <- Sys.getenv("NOAA_KEY")
+  result <- anlz_hydroload(2019, noaa_key)
+
+  expect_is(result, 'data.frame')
+
 })
