+++ tests/testthat/helper-data.R
@@ -1,0 +1,4 @@
+library(tbeptools)
+
+# data for radar plots
+cntdat <- anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2018, radar = TRUE)
