+++ NAMESPACE
@@ -64,6 +64,10 @@ importFrom(RCurl,curlOptions)
 importFrom(RCurl,getURL)
 importFrom(dplyr,"%>%")
 importFrom(grDevices,rgb)
+importFrom(graphics,arrows)
+importFrom(graphics,points)
+importFrom(graphics,polygon)
+importFrom(graphics,text)
 importFrom(lubridate,month)
 importFrom(methods,.S4methods)
 importFrom(methods,new)
@@ -75,6 +79,7 @@ importFrom(reactable,colDef)
 importFrom(stats,ecdf)
 importFrom(stats,median)
 importFrom(stats,na.omit)
+importFrom(stats,pbinom)
 importFrom(stats,sd)
 importFrom(tools,md5sum)
 importFrom(utils,download.file)
