+++ .github/workflows/test-coverage.yaml
@@ -15,6 +15,7 @@ jobs:
     runs-on: macOS-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
+      NOAA_KEY: ${{ secrets.NOAA_KEY}}
     steps:
       - uses: actions/checkout@v2
 
