+++ .github/workflows/R-CMD-check.yaml
@@ -31,6 +31,7 @@ jobs:
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
       RSPM: ${{ matrix.config.rspm }}
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
+      NOAA_KEY: ${{ secrets.NOAA_KEY}}
 
     steps:
       - uses: actions/checkout@v2
