### 2021-06-23 b5ad76e
cleaning up checks again
