+++ tests/testthat/test-anlz_tdlcrkindic.R
@@ -25,7 +25,7 @@ test_that("Checking anlz_tdlcrkindic radar plot output", {
                                       JEI = c("SC37", "SC37", "SC37", "SC37", "SC37", "SC37"),
                                       wbid = c("2026", "2026", "2026", "2026", "2026", "2026"),
                                       var = c("ch_tn_rat_ind", "chla_ind", "do_prop", "nox_ind", "tn_ind", "tsi_ind"),
-                                      val = c(0, 0, 1, NaN, 0, 0)), row.names = c(NA, -6L), class = c("tbl_df", "tbl", "data.frame")))
+                                      val = c(0, 0, 100, NaN, 0, 0)), row.names = c(NA, -6L), class = c("tbl_df", "tbl", "data.frame")))
 
 })
 
