### 2021-06-23 1b2cfe2
fixing build again, ncdc maybe works with NULL key
