+++ .github/workflows/pkgdown.yaml
@@ -52,7 +52,7 @@ jobs:
         run: |
           remotes::install_deps(dependencies = TRUE)
           remotes::install_github("fawda123/Jabbrev")
-          install.packages(c("ggplot2", "patchwork", "plotly", "rnoaa", "pkgdown"))
+          install.packages(c("ggplot2", "patchwork", "plotly", "rnoaa", "cachem", "pkgdown"))
         shell: Rscript {0}
 
       - name: Install package
