### 2021-06-29 0adde9b
update installation instructions for r-universe
