### 2021-07-02 e7e229f
added additional sizing arguments for show_tdlcrkradar
