### 2021-07-02 0d54294
fix typo in radar plot label
