### 2021-07-02 e0b2da1
removed extrafont depends, removed zzz.R, removed inst folder
