+++ R/tidalcreeks.R
@@ -43,6 +43,6 @@
 #'   select(id, wbid, JEI, name, class, Creek_Length_m)
 #'
 #' # save
-#' save(tidalcreeks, file = here::here('data', 'tidalcreeks.RData'), compress = 'xz')
+#' save(tidalcreeks, file = 'data/tidalcreeks.RData', compress = 'xz')
 #' }
 "tidalcreeks"
