+++ vignettes/tbni.Rmd
@@ -22,7 +22,6 @@ knitr::opts_chunk$set(
 
 # libraries
 library(tbeptools) 
-library(here)
 library(bookdown)
 library(patchwork)
 library(dplyr)
@@ -33,9 +32,7 @@ library(sf)
 
 st_crs(fimstations) <- 4326
 
-# Jabbrev::bib_scrp(here('vignettes', 'tbni.Rmd'), bib_new = here('vignettes', 'tbnirefs.bib'), ext_bib = 'https://raw.githubusercontent.com/tbep-tech/tbep-refs/master/bibs.bib')
-
-# spelling::spell_check_files(here("vignettes", "tbni.Rmd"))
+# spelling::spell_check_files("vignettes/tbni.Rmd")
 ```
 
 ## Background
