+++ vignettes/tbbi.Rmd
@@ -21,8 +21,7 @@ knitr::opts_chunk$set(
 )
 
 # libraries
-library(tbeptools) 
-library(here)
+library(tbeptools)
 library(bookdown)
 library(patchwork)
 library(dplyr)
@@ -32,9 +31,7 @@ library(ggplot2)
 library(sf)
 library(tibble)
 
-# Jabbrev::bib_scrp(here('vignettes', 'tbbi.Rmd'), bib_new = here('vignettes', 'tbbirefs.bib'), ext_bib = 'https://raw.githubusercontent.com/tbep-tech/tbep-refs/master/bibs.bib')
-
-# spelling::spell_check_files(here::here("vignettes", "tbbi.Rmd"))
+# spelling::spell_check_files("vignettes/tbbi.Rmd")
 ```
 
 ## Background
