+++ .github/workflows/R-CMD-check.yaml
@@ -74,7 +74,6 @@ jobs:
       - name: Install dependencies
         run: |
           remotes::install_deps(dependencies = TRUE)
-          remotes::install_github("fawda123/Jabbrev")
           remotes::install_cran("rcmdcheck")
         shell: Rscript {0}
 
