+++ .github/workflows/test-coverage.yaml
@@ -46,7 +46,6 @@ jobs:
         run: |
           install.packages(c("remotes"))
           remotes::install_deps(dependencies = TRUE)
-          remotes::install_github("fawda123/Jabbrev")
           remotes::install_cran("covr")
         shell: Rscript {0}
 
