### 2021-07-06 77a88df
correct install instructions
