+++ README.md
@@ -16,7 +16,15 @@ R package for Tampa Bay Estuary Program functions. Please see the [vignette](htt
 The package can be installed from [r-universe](https://tbep-tech.r-universe.dev).  The source code is available on the tbep-tech GitHub group web page: <https://github.com/tbep-tech/tbeptools>.  Note that tbeptools only needs to be installed once, but it needs to be loaded every new R session (i.e., `library(tbeptools)`).
 
 ```r
-install.packages('tbeptools', repos = 'https://tbep-tech/r-universe.dev')
+# enable repos
+options(repos = c(
+    fawda123 = 'https://fawda123.r-universe.dev',
+    CRAN = 'https://cloud.r-project.org'))
+
+# install tbeptools
+install.packages('tbeptools')
+
+# load tbeptools
 library(tbeptools)
 ```
 
