+++ vignettes/tidalrefs.bib
@@ -10,14 +10,26 @@
 	misc={},
 	url={https://drive.google.com/file/d/1YeaZUqIFiak8KdJLSqNrCsQETf8_XTzZ/view?usp=drivesdk},
 }
-@techreport{tbep1319,
-	author={{FFWCC (K.E. Flaherty-Walia, B.J. Williams, B. Barbara, B.L. Winner, P.W. Stevens, T.C. MacDonald, T.S. Switzer, and S.F. Keenan)}},
-	year={2019},
-	title={{Fish Communities Associated with Hard Bottom Habitats in Tampa Bayard Bottom Habitats in Tampa Bay}},
-	number={13-19},
+@techreport{tbep0220,
+	author={{Janicki Environmental and Mote Marine Laboratory}},
+	year={2020},
+	title={{Southwest Florida Tidal Creeks: Nutrient Management Framework and Indicator Development}},
+	number={02-20},
 	publisher={Tampa Bay Estuary Program},
 	address={St. Petersburg, Florida},
-	pages={53},
+	pages={},
 	misc={},
-	url={https://drive.google.com/file/d/1-uG8rSQC7zLXlmcJjSp7B79ZWdl1Tq3E/view?usp=sharing},
+	url={https://drive.google.com/file/d/1n4okSVtNSswVXv1Lje-qMopNZdvUrQnt/view?usp=drivesdk},
+}
+@article{tbep1121,
+	author={{M.R. Wessel, J.R. Leverone, M.W. Beck, E.T. Sherwood, J. Hecker, S. West, A. Janicki}},
+	year={2021},
+	title={{Developing a water quality assessment framework for southwest Florida tidal creeks}},
+	journal={Estuaries & Coasts},
+	volume={},
+	number={},
+	pages={},
+	doi={https://doi.org/10.1007/s12237-021-00974-7},
+	misc={},
+	url={https://doi.org/10.1007/s12237-021-00974-7},
 }
