### 2021-07-10 554a29b
cleaned up old code and added code to remove fdep qualifiers. struggled with best way to do this.Also added run creek61.rmd to help guide you through inputting new data for update. i dont think this actually calls the new data but you have names for it etc. Data are on ftp site. will email link
