### 2021-07-11 1d066e4
fixing docs/tests and removing files for build
