+++ run creek61.Rmd
@@ -1,67 +1,0 @@
----
-title: "Tidal Creek Assessment"
-csl: stylefile.csl
-bibliography: tidalrefs.bib
-date: "`r Sys.Date()`"
-output: 
-  rmarkdown::html_vignette:
-    toc: true
-vignette: >
-  %\VignetteIndexEntry{Tidal Creek Assessment}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-```{r setup, include = FALSE}
-
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>",
-  message = F, warning = F, 
-  fig.align = 'center', 
-  cache = F
-)
-
-# libraries
-library(tbeptools) 
-library(bookdown)
-library(dplyr)
-library(knitr)
-library(mapview)
-library(plotly)
-library(haven)
-library(sf)
-
-sf::st_crs(tidalcreeks) <- 4326
-# spelling::spell_check_files("vignettes/tidalcreeks.Rmd")
-```
-
-
-```{r}
-
-setwd("E:/To Marcus/Feed dashboard")
-
-#iwrraw<-read_sas("iwrraw_creeks61.sas7bdat")
-#creek_pop<-read_sas("creek_pop61.sas7bdat")
-
-#save(iwrraw,file="iwrraw.rdata")
-#save(creek_pop,file="creek_pop.rdata")
-
-Creek_jei<- sf::read_sf(dsn = ".", layer = "TidalCreek_All_Line") 
-
-```
-
-## The Dashboard functions run using code below after loading iwrraw into the working environment but im not sure that this isnt going back to the old file? Not sure where iwrraw is stored in the main repo. 
-
-never the less, these answers wont be the same because they use the old creek pop file and the  old line coverage where in some cases the line file was improved and creek length may have changed. 
-
-```{r}
-results <- anlz_tdlcrk(tidalcreeks, iwrraw)
-results
-```
-
-
-```{r, eval = T, cache = F}
-show_tdlcrk(results)
-```
-
