### 2021-07-12 f8f9e0f
fixing some bugs in show tidal creek functions for iwr update, fixing tests for tidal creeks for iwr update
