+++ tests/testthat/test-anlz_iwrraw.R
@@ -1,9 +1,9 @@
 test_that("Checking anlz_iwrraw", {
 
-  result <- anlz_iwrraw(iwrraw, tidalcreeks, yr = 2018) %>%
+  result <- anlz_iwrraw(iwrraw, tidalcreeks, yr = 2021) %>%
     pull(masterCode) %>%
     unique
   expect_equal(result, c("COND", "SALIN", "COLOR", "DO", "DOSAT", "NO23", "ORGN", "TN",
-                         "TP", "CHLAC", "TKN", "TSS", "TURB"))
+                         "TP", "CHLAC", "TKN", "TURB", "TSS"))
 
 })
