+++ tests/testthat/test-show_tdlcrkradar.R
@@ -1,12 +1,12 @@
 test_that("Checking show_tdlcrkradar class", {
 
-  result <- show_tdlcrkradar(494, cntdat)
+  result <- show_tdlcrkradar(494, cntdatrdr)
   expect_is(result, 'NULL')
 
 })
 
 test_that("Checking show_tdlcrkradar no data", {
 
-  expect_error(show_tdlcrkradar(495, cntdat))
+  expect_error(show_tdlcrkradar(495, cntdatrdr))
 
 })
