+++ R/anlz_iwrraw.R
@@ -2,7 +2,7 @@
 #'
 #' Format raw IWR data
 #'
-#' @param iwrraw FDEP impaired waters rule run 56 data base as \code{\link{data.frame}}
+#' @param iwrraw FDEP impaired waters rule data base as \code{\link{data.frame}}
 #' @param tidalcreeks \code{\link[sf]{sf}} object for population of tidal creeks
 #' @param yr numeric for reference year to evaluate, scores are based on the planning period beginning ten years prior to this date
 #'
