+++ tests/testthat/test-show_tdlcrkindiccdf.R
@@ -1,11 +1,9 @@
 test_that("Checking show_tdlcrkindiccdf class", {
-  cntdat <- anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2018)
-  result <- show_tdlcrkindiccdf(197, cntdat, thrsel = TRUE)
+  result <- show_tdlcrkindiccdf(479, cntdat, thrsel = TRUE, yr = 2021)
   expect_is(result, 'plotly')
 })
 test_that("Checking show_tdlcrkindiccdf empty data", {
-  cntdat <- anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2018)
-  result <- show_tdlcrkindiccdf('xxxxxxx', cntdat, thrsel = TRUE)
+  result <- show_tdlcrkindiccdf('xxxxxxx', cntdat, thrsel = TRUE, yr = 2021)
   expect_null(result)
 })
 
