+++ tests/testthat/test-show_tdlcrkmatrix.R
@@ -1,11 +1,10 @@
-dat <- anlz_tdlcrk(tidalcreeks, iwrraw, yr = 2018)
 test_that("Checking show_tdlcrkmatrix class", {
-  result <- show_tdlcrkmatrix(dat)
+  result <- show_tdlcrkmatrix(tdldat)
   expect_is(result, 'ggplot')
 })
 test_that("Checking show_tdlcrkmatrix class argument", {
-  expect_error(show_tdlcrkmatrix(dat, class = 'asdf'))
+  expect_error(show_tdlcrkmatrix(tdldat, class = 'asdf'))
 })
 test_that("Checking show_tdlcrkmatrix score argument", {
-  expect_error(show_tdlcrkmatrix(dat, score = 'asdf'))
+  expect_error(show_tdlcrkmatrix(tdldat, score = 'asdf'))
 })
