+++ R/anlz_tdlcrkindic.R
@@ -3,7 +3,7 @@
 #' Estimate tidal creek water quality indicators to support report card scores
 #'
 #' @param tidalcreeks \code{\link[sf]{sf}} object for population of tidal creeks
-#' @param iwrraw FDEP impaired waters rule run 56 data base as \code{\link{data.frame}}
+#' @param iwrraw FDEP impaired waters rule data base as \code{\link{data.frame}}
 #' @param yr numeric for reference year to evaluate, scores are based on the planning period beginning ten years prior to this date
 #' @param radar logical indicating if output is for \code{\link{show_tdlcrkradar}}, see details
 #'
@@ -21,9 +21,9 @@
 #' @concept analyze
 #'
 #' @examples
-#' dat <- anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2018)
+#' dat <- anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2021)
 #' head(dat)
-anlz_tdlcrkindic <- function(tidalcreeks, iwrraw, yr = 2018, radar = FALSE) {
+anlz_tdlcrkindic <- function(tidalcreeks, iwrraw, yr = 2021, radar = FALSE) {
 
   # format iwr data
   iwrdat <- anlz_iwrraw(iwrraw, tidalcreeks, yr = yr)
