+++ man/show_tdlcrkradar.Rd
@@ -41,7 +41,7 @@ See details in \code{\link{anlz_tdlcrkindic}} for an explanation of the indicato
 Internal code borrowed heavily from \code{\link[fmsb]{radarchart}}.
 }
 \examples{
-cntdat <- anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2018, radar = TRUE)
+cntdat <- anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2021, radar = TRUE)
 show_tdlcrkradar(35, cntdat)
 }
 \concept{show}

[…]
