### 2021-07-12 6f19a1a
minor doc edits
