### 2021-07-12 #3243 @kthyng
Hi all, I'm not sure what happens in summer but time is evaporating. @richardsc it sounds like you'll be back from vacation soon so I will ask that you take this on as your inaugural JOSS review! Thank you @padilla410 for your offer, and if you'd like to be a reviewer in the future you can sign up [here](https://joss.theoj.org/reviewer-signup.html).
