### 2021-07-12 #3485 @whedon
[REVIEW]: tbeptools: An R package for synthesizing estuarine data for environmental research
**Submitting author:** @fawda123 (<a href="http://orcid.org/0000-0002-4996-0059">Marcus Beck</a>)
**Repository:** <a href="https://github.com/tbep-tech/tbeptools" target ="_blank">https://github.com/tbep-tech/tbeptools</a>
**Version:** v1.1.0
