### 2021-07-23 259ec39
transect training data returns assessment year
