### 2021-07-23 8250f9f
yr arg to show_compplot
