### 2021-07-23 1afeb8e
fixing tests for show_compplot
