### 2021-07-26 tbep-tech/tbeptools#43 @paleolimbot
[JOSS Review] Clarify statement of need in README and in paper
Hi!

This package is an excellent example of packaging up code that can be well tested to back a Shiny app or other type of automated reporting (e.g., Rmd report). You've backed it up with integration testing with impressive coverage. I didn't see any reference to the Shiny app in the README or the paper and I think that's a shame...the Statement of Need in my opinion is that this is a good example of a well-designed reporting system to communicate environmental data to clients and stakeholders. A short paragraph with a link might work well for the README...in the paper I would love to see an all-out flowchart/screenshot of the Shiny app. I know the Shiny app isn't your submission but the design of the whole thing is where the value is for your readers (very few of whom actually need data from the Tampa Bay Estuary but might be interested in doing this kind of thing with their own monitoring program).

(Linking https://github.com/openjournals/joss-reviews/issues/3485 )
