### 2021-07-26 tbep-tech/tbeptools#44 @paleolimbot
[JOSS Review] Widen the scope of literature review in paper
Hi again!

In your paper submission you've noted one other package that reports estuarine data. I think there's two more broad types of packages that would be useful for readers of your paper to know about: (1) general water quality data access (like dataRetrieval, which I think is in your `Imports`) and (2) calculation of policy thresholds in an open and repeatable way (like https://github.com/bcgov/wqindex ). This gives you an opportunity to comment on how you solved the issues associated with each of these, such as the download/cache system you need for a data access package (or as package data? or I think you have some of both here?). For policy thresholds (here, the difference between 'Stay the course', 'Caution', and 'On alert'), I imagine it gave you an opportunity to document how these are calculated and allow clients/stakeholders to reproduce your work.

(Linking: https://github.com/openjournals/joss-reviews/issues/3485 )
