### 2021-07-26 tbep-tech/tbeptools#45 @paleolimbot
[JOSS Review] Testing on MacOS?
It looks like you're skipping MacOS on your GitHub Actions...I'm guessing because of the latest issues with sf building. If that is the case, you can conditionally make GitHub actions install the binary version of sf until the build issue has been solved (see https://github.com/paleolimbot/wk/blob/master/.github/workflows/R-CMD-check.yaml#L69-L70 ). I know this is just temporary but I do think it's important to make sure your package will work on MacOS + Windows + Linux.

(Linking https://github.com/openjournals/joss-reviews/issues/3485 )
