### 2021-07-26 tbep-tech/tbeptools#46 @paleolimbot
[JOSS Review] Add brief example usage to the paper
I was surprised that there wasn't some example usage in your paper...you've got an [excellent vignette with great material](https://tbep-tech.github.io/tbeptools/articles/intro.html)...I think a quick demo of a `read_*()` then `anlz_*()` then `show_*()` would be useful in the paper (and in your README to give users an idea of where to start with your package).

(Linking https://github.com/openjournals/joss-reviews/issues/3485 )
