### 2021-07-26 tbep-tech/tbeptools#43 @paleolimbot
(Just clarifying that this also applies to the Summary section of the paper...as in, your summary should be revised to reflect your updated statement of need)
