### 2021-07-26 #3485 @paleolimbot
My first round of comments are in as issues in the target repo! It's an excellent example of well-tested code wrapped up as a package to support automated reporting of environmental data and will be an excellent contribution to JOSS!

The three things I'd like to see are (1) a tiny bit of example usage in the paper, (2) updating the Statement of Need section to mention the Shiny app and the role that this package fills in the reporting component, and (3) a slightly broader scope when reviewing previous work. Ultimately I'd like the paper to speak to the bigger problems that other research software writers need to solve outside the scope of the Tampa Bay Estuary...those three bullet points are my two cents on what that is although I'm happy to entertain other ideas!
