+++ R/globalVariables.R
@@ -36,7 +36,8 @@ globalVariables(c("Chlorophyll_a uncorr_Q", "Latitude", "Longitude", "SampleTime
                   'All', 'All (wt)', 'cntwts', 'grndn', 'nwts', 'per', 'tbbicatid', 'tbbicatwt', 'total',
                   'AreaAbbr', 'AreaID', 'FundingId', 'FundingProject', 'no23_source', 'no23_tidal', 'DOSAT',
                   'DOSAT_prop', 'ch_tn_rat_ind', 'chla_ind', 'do_bnml', 'do_cnt', 'do_prop', 'fo', 'nox_ind',
-                  'sta', 'tn_ind', 'tsi_ind', 'xcd'))
+                  'sta', 'tn_ind', 'tsi_ind', 'xcd', 'AssessmentYear', 'disqual', 'disqual2', 'newComment',
+                  'rCode'))
 
 #' @importFrom grDevices rgb
 NULL
