+++ R/fimdata.R
@@ -1,15 +1,16 @@
-#' FIM data for Tampa Bay Nekton Index current as of 04242020
+#' FIM data for Tampa Bay Nekton Index current as of 07262021
 #'
-#' FIM data for Tampa Bay Nekton Index current as of 04242020
+#' FIM data for Tampa Bay Nekton Index current as of 07262021
 #'
-#' @format A data frame with 41653 rows and 17 variables:
+#' @format A data frame with 61148 rows and 19 variables:
 #' \describe{
 #'   \item{Reference}{chr}
 #'   \item{Sampling_Date}{Date}
+#'   \item{Latitude}{num}
+#'   \item{Longitude}{num}
 #'   \item{Zone}{chr}
 #'   \item{Grid}{int}
 #'   \item{NODCCODE}{chr}
-#'   \item{bay_segment}{chr}
 #'   \item{Year}{num}
 #'   \item{Month}{num}
 #'   \item{Total_N}{num}
@@ -21,6 +22,7 @@
 #'   \item{Feeding_Cat}{chr}
 #'   \item{Feeding_Guild}{chr}
 #'   \item{Selected_Taxa}{chr}
+#'   \item{bay_segment}{chr}
 #'   }
 #'
 #' @concept data
