+++ man/fimstations.Rd
@@ -5,7 +5,7 @@
 \alias{fimstations}
 \title{Spatial data object of FIM stations including Tampa Bay segments}
 \format{
-A simple features \code{\link[sf]{sf}} object (POINT) with 6186 features and 4 fields, +proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs
+A simple features \code{\link[sf]{sf}} object (POINT) with 6796 features and 2 fields, +proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs
 \describe{
   \item{Reference}{num}
   \item{bay_segment}{chr}
