+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: tbeptools
 Title: Data and Indicators for the Tampa Bay Estuary Program
 Version: 1.0.2
-Date: 2021-07-23
+Date: 2021-07-26
 Authors@R: c(
     person(given = "Marcus",
       family = "Beck",
