### 2021-07-26 afa702b
inner join on clipped stations by segment to fimdata
