+++ R/fimdata.R
@@ -2,7 +2,7 @@
 #'
 #' FIM data for Tampa Bay Nekton Index current as of 07262021
 #'
-#' @format A data frame with 61148 rows and 19 variables:
+#' @format A data frame with 45945 rows and 19 variables:
 #' \describe{
 #'   \item{Reference}{chr}
 #'   \item{Sampling_Date}{Date}
