+++ tests/testthat/test-anlz_tbniscr.R
@@ -2,12 +2,12 @@ test_that("Checking anlz_tbniscr", {
 
   tbniscr <- anlz_tbniscr(fimdata)
 
-  result <- tbniscr[nrow(tbniscr), ]
-  expect_equal(result, structure(list(Reference = "TBM2019121309", Year = 2019, Month = 12,
-                                      Season = "Winter", bay_segment = "MTB", TBNI_Score = 38,
-                                      NumTaxa = 5, ScoreNumTaxa = 5, BenthicTaxa = 5, ScoreBenthicTaxa = 5,
+  result <- tbniscr[1, ]
+  expect_equal(result, structure(list(Reference = "TBM1998010906", Year = 1998, Month = 1,
+                                      Season = "Winter", bay_segment = "OTB", TBNI_Score = 18,
+                                      NumTaxa = 2, ScoreNumTaxa = 2, BenthicTaxa = 0, ScoreBenthicTaxa = 0,
                                       TaxaSelect = 0, ScoreTaxaSelect = 0, NumGuilds = 2, ScoreNumGuilds = 5,
-                                      Shannon = 0.727591345714938, ScoreShannon = 4), row.names = c(NA,
+                                      Shannon = 0.362210557135449, ScoreShannon = 2), row.names = c(NA,
                                                                                                     -1L), class = c("tbl_df", "tbl", "data.frame"))
   )
 
@@ -17,11 +17,11 @@ test_that("Checking anlz_tbniscr, raw = F", {
 
   tbniscr <- anlz_tbniscr(fimdata, raw = F)
 
-  result <- tbniscr[nrow(tbniscr), ]
-  expect_equal(result, structure(list(Reference = "TBM2019121309", Year = 2019, Month = 12,
-                                      Season = "Winter", bay_segment = "MTB", TBNI_Score = 38,
-                                      ScoreNumTaxa = 5, ScoreBenthicTaxa = 5, ScoreTaxaSelect = 0,
-                                      ScoreNumGuilds = 5, ScoreShannon = 4), row.names = c(NA,
+  result <- tbniscr[1, ]
+  expect_equal(result, structure(list(Reference = "TBM1998010906", Year = 1998, Month = 1,
+                                      Season = "Winter", bay_segment = "OTB", TBNI_Score = 18,
+                                      ScoreNumTaxa = 2, ScoreBenthicTaxa = 0, ScoreTaxaSelect = 0,
+                                      ScoreNumGuilds = 5, ScoreShannon = 2), row.names = c(NA,
                                                                                            -1L), class = c("tbl_df", "tbl", "data.frame"))
   )
 
