### 2021-07-27 da1c41a
fix to segment assignments for fimstations
