+++ tests/testthat/test-anlz_tdlcrk.R
@@ -5,7 +5,7 @@ test_that("Checking anlz_tldcrk", {
 
   expect_equal(result, c("Monitor", "No Data", "No Data", "Monitor", "No Data", "Monitor",
                          "Investigate", "Monitor", "No Data", "No Data", "No Data", "No Data",
-                         "No Data", "No Data", "No Data", "Investigate", "No Data", "No Data",
+                         "No Data", "No Data", "No Data", "Monitor", "No Data", "No Data",
                          "No Data", "No Data", "No Data", "No Data", "No Data", "No Data",
                          "No Data", "Monitor", "No Data", "Monitor", "No Data", "No Data",
                          "No Data", "No Data", "No Data", "Monitor", "No Data", "No Data",
@@ -14,7 +14,7 @@ test_that("Checking anlz_tldcrk", {
                          "No Data", "No Data", "No Data", "No Data", "No Data", "Monitor",
                          "No Data", "No Data", "No Data", "Monitor", "Monitor", "No Data",
                          "No Data", "Monitor", "No Data", "No Data", "Prioritize", "No Data",
-                         "No Data", "Monitor", "No Data", "Monitor", "No Data", "No Data",
+                         "No Data", "No Data", "No Data", "Monitor", "No Data", "No Data",
                          "Monitor", "No Data", "No Data", "No Data", "Monitor", "No Data",
                          "No Data", "Monitor", "Caution", "No Data", "No Data", "No Data",
                          "No Data", "No Data", "Monitor", "No Data", "Monitor", "No Data",
@@ -24,47 +24,45 @@ test_that("Checking anlz_tldcrk", {
                          "No Data", "No Data", "No Data", "No Data", "No Data", "No Data",
                          "No Data", "No Data", "No Data", "No Data", "No Data", "No Data",
                          "No Data", "No Data", "No Data", "No Data", "No Data", "No Data",
-                         "No Data", "Monitor", "No Data", "No Data", "No Data", "No Data",
+                         "No Data", "No Data", "No Data", "No Data", "No Data", "No Data",
                          "No Data", "No Data", "No Data", "No Data", "No Data", "Monitor",
-                         "Investigate", "Prioritize", "Prioritize", "No Data", "Investigate",
-                         "No Data", "No Data", "No Data", "No Data", "No Data", "Prioritize",
-                         "No Data", "Investigate", "No Data", "No Data", "No Data", "Prioritize",
-                         "Monitor", "Monitor", "Monitor", "No Data", "No Data", "Monitor",
-                         "No Data", "No Data", "No Data", "Monitor", "Monitor", "Monitor",
-                         "Monitor", "No Data", "No Data", "Caution", "No Data", "No Data",
-                         "Monitor", "Monitor", "Monitor", "No Data", "No Data", "Monitor",
-                         "No Data", "Monitor", "No Data", "Monitor", "Prioritize", "Monitor",
-                         "Prioritize", "Prioritize", "No Data", "No Data", "No Data",
-                         "Monitor", "No Data", "No Data", "No Data", "No Data", "Monitor",
-                         "No Data", "No Data", "Monitor", "No Data", "No Data", "Prioritize",
-                         "Monitor", "Monitor", "Prioritize", "Monitor", "Monitor", "Prioritize",
+                         "No Data", "Prioritize", "No Data", "No Data", "No Data", "No Data",
+                         "No Data", "No Data", "No Data", "Monitor", "No Data", "No Data",
+                         "No Data", "No Data", "Monitor", "No Data", "Monitor", "No Data",
+                         "No Data", "Monitor", "No Data", "No Data", "No Data", "Monitor",
+                         "Monitor", "Monitor", "Monitor", "No Data", "No Data", "Caution",
[…]
