+++ tests/testthat/test-show_tdlcrkindic.R
@@ -1,5 +1,5 @@
 test_that("Checking show_tdlcrkindic class", {
-  result <- show_tdlcrkindic(479, cntdat, thrsel = TRUE, yr = 2021)
+  result <- show_tdlcrkindic(1, cntdat, thrsel = TRUE, yr = 2021)
   expect_is(result, 'plotly')
 })
 test_that("Checking show_tdlcrkindic empty data", {
