+++ tests/testthat/test-anlz_iwrraw.R
@@ -1,9 +1,12 @@
-test_that("Checking anlz_iwrraw", {
+test_that("Checking correct parameter codes in anlz_iwrraw", {
 
   result <- anlz_iwrraw(iwrraw, tidalcreeks, yr = 2021) %>%
     pull(masterCode) %>%
     unique
-  expect_equal(result, c("COND", "SALIN", "COLOR", "DO", "DOSAT", "NO23", "ORGN", "TN",
-                         "TP", "CHLAC", "TKN", "TURB", "TSS"))
+
+  chk <- any(!result %in% c("COND", "SALIN", "COLOR", "DO", "DOSAT", "NO23", "ORGN", "TN",
+                           "TP", "CHLAC", "TKN", "TURB", "TSS"))
+
+  expect_false(chk)
 
 })
