+++ tests/testthat/test-show_tdlcrkindiccdf.R
@@ -1,5 +1,5 @@
 test_that("Checking show_tdlcrkindiccdf class", {
-  result <- show_tdlcrkindiccdf(479, cntdat, thrsel = TRUE, yr = 2021)
+  result <- show_tdlcrkindiccdf(1, cntdat, thrsel = TRUE, yr = 2021)
   expect_is(result, 'plotly')
 })
 test_that("Checking show_tdlcrkindiccdf empty data", {
