### 2021-08-01 3d4e65f
fixing tidal creek examples
