+++ R/show_tdlcrkradar.R
@@ -22,7 +22,7 @@
 #'
 #' @examples
 #' cntdat <- anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2021, radar = TRUE)
-#' show_tdlcrkradar(494, cntdat)
+#' show_tdlcrkradar(495, cntdat)
 show_tdlcrkradar <- function (id, cntdat, col = '#338080E6', ptsz = 1, lbsz = 0.8, valsz = 1, brdwd = 5){
 
   axistype <- 3
