+++ man/show_tdlcrkindiccdf.Rd
@@ -34,6 +34,6 @@ This function returns several empirical cumulative distribution plots for the ti
 }
 \examples{
 cntdat <- anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2021)
-show_tdlcrkindiccdf(479, cntdat, thrsel = TRUE)
+show_tdlcrkindiccdf(495, cntdat, thrsel = TRUE)
 }
 \concept{show}
