+++ tests/testthat/test-anlz_tbbimed.R
@@ -1,11 +1,11 @@
 test_that("Checking results for anlz_tbbimed", {
-  tbbiscr <- anlz_tbbiscr(benthicdata)
+
   results <- anlz_tbbimed(tbbiscr)
-  results <- results[nrow(results), ]
-  expect_equal(results, structure(list(bay_segment = structure(5L, .Label = c("HB", "OTB",
+  results <- results[1, ]
+  expect_equal(results, structure(list(bay_segment = structure(8L, .Label = c("HB", "OTB",
                                                                               "MTB", "LTB", "TCB", "MR", "BCB", "All", "All (wt)"), class = "factor"),
-                                       yr = 2018, Degraded = 0, Healthy = 0.666666666666667, Intermediate = 0.333333333333333,
-                                       TBBICat = structure(3L, .Label = c("Poor", "Fair", "Good"
+                                       yr = 1993, Degraded = 0.3, Healthy = 0.244444444444444, Intermediate = 0.455555555555556,
+                                       TBBICat = structure(1L, .Label = c("Poor", "Fair", "Good"
                                        ), class = "factor")), row.names = c(NA, -1L), groups = structure(list(
                                          .rows = structure(list(1L), ptype = integer(0), class = c("vctrs_list_of",
                                                                                                    "vctrs_vctr", "list"))), row.names = c(NA, -1L), class = c("tbl_df",
@@ -14,13 +14,13 @@ test_that("Checking results for anlz_tbbimed", {
 })
 
 test_that("Checking results for anlz_tbbimed, rev = T", {
-  tbbiscr <- anlz_tbbiscr(benthicdata)
+
   results <- anlz_tbbimed(tbbiscr, rev = T)
-  results <- results[nrow(results), ]
-  expect_equal(results, structure(list(bay_segment = structure(5L, .Label = c("All (wt)",
+  results <- results[1, ]
+  expect_equal(results, structure(list(bay_segment = structure(2L, .Label = c("All (wt)",
                                                                               "All", "BCB", "MR", "TCB", "LTB", "MTB", "OTB", "HB"), class = "factor"),
-                                       yr = 2018, Degraded = 0, Healthy = 0.666666666666667, Intermediate = 0.333333333333333,
-                                       TBBICat = structure(3L, .Label = c("Poor", "Fair", "Good"
+                                       yr = 1993, Degraded = 0.3, Healthy = 0.244444444444444, Intermediate = 0.455555555555556,
+                                       TBBICat = structure(1L, .Label = c("Poor", "Fair", "Good"
                                        ), class = "factor")), row.names = c(NA, -1L), groups = structure(list(
                                          .rows = structure(list(1L), ptype = integer(0), class = c("vctrs_list_of",
                                                                                                    "vctrs_vctr", "list"))), row.names = c(NA, -1L), class = c("tbl_df",
