+++ R/benthicdata.R
@@ -1,6 +1,6 @@
-#' Benthic data for the Tampa Bay Benthic Index current as of 10222020
+#' Benthic data for the Tampa Bay Benthic Index current as of 08012021
 #'
-#' Benthic data for the Tampa Bay Benthic Index current as of 10222020
+#' Benthic data for the Tampa Bay Benthic Index current as of 08012021
 #'
 #' @format A nested \code{\link[tibble]{tibble}} with 3 rows and 2 variables:
 #' \describe{
