+++ tests/testthat/helper-data.R
@@ -8,3 +8,6 @@ cntdat <- anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2021)
 
 # tidal creek scores
 tdldat <- anlz_tdlcrk(tidalcreeks, iwrraw, yr = 2021)
+
+# benthic index scores
+tbbiscr <- anlz_tbbiscr(benthicdata)
