+++ R/show_tbbimatrix.R
@@ -25,7 +25,7 @@
 #' @examples
 #' tbbiscr <- anlz_tbbiscr(benthicdata)
 #' show_tbbimatrix(tbbiscr)
-show_tbbimatrix <- function(tbbiscr, bay_segment = c('HB', 'OTB', 'MTB', 'LTB', 'TCB', 'MR', 'BCB', 'All', 'All (wt)'), yrrng = c(1993, 2018), alph = 1, txtsz = 3, family = NA, rev = FALSE, position = 'top', plotly = FALSE){
+show_tbbimatrix <- function(tbbiscr, bay_segment = c('HB', 'OTB', 'MTB', 'LTB', 'TCB', 'MR', 'BCB', 'All', 'All (wt)'), yrrng = c(1993, 2019), alph = 1, txtsz = 3, family = NA, rev = FALSE, position = 'top', plotly = FALSE){
 
   # annual average by segment
   toplo <- anlz_tbbimed(tbbiscr, bay_segment, rev = rev, yrrng = yrrng)
