+++ man/show_tbbimatrix.Rd
@@ -7,7 +7,7 @@
 show_tbbimatrix(
   tbbiscr,
   bay_segment = c("HB", "OTB", "MTB", "LTB", "TCB", "MR", "BCB", "All", "All (wt)"),
-  yrrng = c(1993, 2018),
+  yrrng = c(1993, 2019),
   alph = 1,
   txtsz = 3,
   family = NA,
