+++ R/read_formbenthic.R
@@ -1,10 +1,10 @@
 #' Format benthic data for the Tampa Bay Benthic Index
 #'
-#' @param channel A \code{\link[RODBC]{RODBC}} class connection object to the .mdb benthic database
+#' @param pathin A path to unzipped csv files with base tables used to calculate benthic index
 #'
 #' @return A nested \code{\link[tibble]{tibble}} of station, field sample, and taxa data
 #'
-#' @details Function is used internally within \code{\link{read_importbenthic}}, see the help file for limitations on using the function outside of Windows.
+#' @details Function is used internally within \code{\link{read_importbenthic}}
 #'
 #' @concept read
 #'
@@ -14,42 +14,50 @@
 #'
 #' @examples
 #' \dontrun{
-#' library(RODBC)
 #'
-#' # setup driver and path to .mdb
-#' drvr <- 'Driver={Microsoft Access Driver (*.mdb, *.accdb)};'
-#' path <- 'C:/Users/mbeck/Desktop/benthic/EPC DataSubmittals.mdb'
-#' odbccall <- paste0(drvr, 'DBQ=', path)
+#' # location to download data
+#' path <- '~/Desktop/benthic.zip'
 #'
-#' # create channel and pass to read_formbenthic
-#' channel <- RODBC::odbcDriverConnect(odbccall)
-#' read_formbenthic(channel)
+#' # download
+#' urlin <- 'ftp://ftp.epchc.org/EPC_ERM_FTP/Benthic_Monitoring/DataDeliverables/BaseTables.zip'
+#' read_dlcurrent(path, download_latest = TRUE, urlin = urlin)
 #'
-#' # close connection
-#' odbcClose(channel)
+#' # unzip
+#' tmppth <- tempfile()
+#' utils::unzip(path, exdir = tmppth, overwrite = TRUE)
+#'
+#' # format benthic data
+#' read_formbenthic(pathin = tmppth)
+#'
+#' # remove temporary path
+#' unlink(tmppth, recursive = TRUE)
+#'
 #' }
-read_formbenthic <- function(channel){
+read_formbenthic <- function(pathin){
 
   tzone <- 'America/Jamaica'
 
   # fetch relevant tables ---------------------------------------------------
 
+  # files
+  fls <- list.files(pathin, recursive = T, full.names = T)
+
   # station metadata
-  Stations <- RODBC::sqlFetch(channel, 'Stations')
-  Header <- RODBC::sqlFetch(channel, 'Header')
-  Programs <- RODBC::sqlFetch(channel, 'Programs')
-  ProgramsStations <- RODBC::sqlFetch(channel, 'ProgramsStations')
-  SegmentName <- RODBC::sqlFetch(channel, 'SegmentName')
-  FundingProject <- RODBC::sqlFetch(channel, 'FundingProject')
+  Stations <- read.csv(grep('/Stations\\.csv$', fls, value = TRUE))
+  Header <- read.csv(grep('Header', fls, value = TRUE))
+  Programs <- read.csv(grep('Programs\\.csv$', fls, value = TRUE))
+  ProgramsStations <- read.csv(grep('ProgramsStations', fls, value = TRUE))
+  SegmentName <- read.csv(grep('SegmentName', fls, value = TRUE))
+  FundingProject <- read.csv(grep('FundingProject', fls, value = TRUE))
 
   # water chem
-  FieldSamples <- RODBC::sqlFetch(channel, 'FieldSamples')
+  FieldSamples <- read.csv(grep('FieldSamples', fls, value = TRUE))
 
   # bug data
[…]
