+++ man/benthicdata.Rd
@@ -20,7 +20,7 @@ Benthic data for the Tampa Bay Benthic Index current as of 08012021
 \examples{
 \dontrun{
 # location to download data
-path <- 'C:/Users/mbeck/Desktop/benthic.zip'
+path <- '~/Desktop/benthic.zip'
 
 # load and assign to object
 benthicdata <- read_importbenthic(path, download_latest = TRUE, remove = TRUE)
