+++ DESCRIPTION
@@ -61,7 +61,6 @@ Imports:
   reactable,
   readxl,
   rnoaa,
-  RODBC,
   sf,
   RCurl,
   tibble,

[…]
