### 2021-08-01 9862fd9
fixing non ascii strings again
