### 2021-08-02 807c738
working on csv import for non ascii to fix ubuntu build
