### 2021-08-16 #3485 @fawda123
@kthyng thanks for checking in, no I have not yet addressed any of the comments from @paleolimbot.  I was waiting for review from @richardsc, but can proceed with revision this wek if preferred.
