### 2021-08-18 9d61691
adding function examples to paper and readme, #46
