+++ paper/paper.Rmd
@@ -44,7 +44,11 @@ csl: apa.csl
 journal: JOSS
 ---
 
+```{r setup, warning = F, message = F, echo = F}
+knitr::opts_chunk$set(message = F, warning = F, echo = T, cache = F, dev.args = list(family = 'serif'), dpi = 400, tidy.opts=list(width.cutoff=80), tidy=TRUE)
 
+options(tibble.width = 60, tibble.print_max = 5, tibble.print_min = 5)
+```
 
 # Summary
 
@@ -60,71 +64,24 @@ The function names were chosen with a typical analysis workflow in mind, where f
 
 The following example demonstrates use of a subset of the functions for water quality data to read a file from the Hillsborough County Environmental Protection Commission long-term monitoring dataset (available from <https://www.tampabay.wateratlas.usf.edu/>), analyze monthly and annual averages by major bay segments of Tampa Bay, and plot an annual time series for one of the bay segments.
 
-
-```r
+```{r thrplotex, fig.height = 3.5, fig.width = 7}
 # load the package
 library(tbeptools)
 
 # read current data
-wqdat <- read_importwq(xlsx = "wqdata.xlsx", download_latest = TRUE)
+wqdat <- read_importwq(xlsx = 'wqdata.xlsx', download_latest = TRUE)
 wqdat
-```
 
-```
-## # A tibble: 26,611 x 22
-##   bay_segment epchc_station SampleTime             yr    mo
-##   <chr>               <dbl> <dttm>              <dbl> <dbl>
-## 1 HB                      6 2021-06-08 10:59:00  2021     6
-## 2 HB                      7 2021-06-08 11:13:00  2021     6
-## 3 HB                      8 2021-06-08 14:15:00  2021     6
-## 4 MTB                     9 2021-06-08 13:14:00  2021     6
-## 5 MTB                    11 2021-06-08 11:30:00  2021     6
-## # ... with 26,606 more rows, and 17 more variables:
-## #   Latitude <dbl>, Longitude <dbl>, Total_Depth_m <dbl>,
-## #   Sample_Depth_m <dbl>, tn <dbl>, tn_q <chr>, sd_m <dbl>,
-## #   sd_raw_m <dbl>, sd_q <chr>, chla <dbl>, chla_q <chr>,
-## #   Sal_Top_ppth <dbl>, Sal_Mid_ppth <dbl>,
-## #   Sal_Bottom_ppth <dbl>, Temp_Water_Top_degC <dbl>,
-## #   Temp_Water_Mid_degC <dbl>, ...
-```
-
-```r
 # analyze monthly and annual means by bay segment
 avedat <- anlz_avedat(wqdat)
 avedat
-```
 
-```
-## $ann
-## # A tibble: 584 x 4
-##      yr bay_segment var         val
-##   <dbl> <chr>       <chr>     <dbl>
-## 1  1974 HB          mean_chla 22.4 
-## 2  1974 LTB         mean_chla  4.24
-## 3  1974 MTB         mean_chla  9.66
-## 4  1974 OTB         mean_chla 10.2 
-## 5  1975 HB          mean_chla 27.9 
-## # ... with 579 more rows
-## 
-## $mos
-## # A tibble: 4,484 x 5
-##   bay_segment    yr    mo var         val
-##   <chr>       <dbl> <dbl> <chr>     <dbl>
-## 1 HB           1974     1 mean_chla 36.2 
-## 2 LTB          1974     1 mean_chla  1.75
-## 3 MTB          1974     1 mean_chla 11.5 
-## 4 OTB          1974     1 mean_chla  4.4 
-## 5 HB           1974     2 mean_chla 42.4 
-## # ... with 4,479 more rows
-```
-
[…]
