### 2021-08-18 caaaa6d
jpeg graphics device
