### 2021-08-18 eb8ef55
thrplot figure to man/figures
