+++ paper/paper.Rmd
@@ -81,6 +81,7 @@ show_thrplot(wqdat, bay_segment = 'HB', yrrng = c(1975, 2020))
 ```
 ```{r, echo = F, results = 'hide'}
 file.remove('wqdata.xlsx')
+file.copy('paper_files/figure-latex/thrplotex-1.jpeg', '../man/figures/')
 ```
 
 Most of the NEPs do not have analysis software to operationalize data import, analysis, and plotting for reporting.  Recently, a similar software package, `peptools` [@Beck21a], was developed for the Peconic Estuary Partnership (New York, USA) using many of the functions in `tbeptools` to develop reporting products for a new water quality monitoring program.  This successful technology transfer demonstrates the added value of presenting these methods in an open source environment available for discovery and reuse by others.  We expect other NEPs to begin using these tools as their application becomes more widespread among estuarine researchers.
