### 2021-08-18 227915b
adding content to paper and readme about dashboards, fix #43
