+++ paper/paper.bib
@@ -14,6 +14,59 @@
 	pages={2}
 }
 
+@software{Beck20a,
+  author       = {Marcus W Beck},
+  title        = {tbep-tech/wq-dash: v1.0},
+  month        = feb,
+  year         = 2020,
+  publisher    = {Zenodo},
+  version      = {v1.0},
+  doi          = {10.5281/zenodo.3648664},
+  url          = {https://doi.org/10.5281/zenodo.3648664}
+}
+
+@software{Beck20b,
+  author       = {Marcus W Beck},
+  title        = {tbep-tech/seagrasstransect-dash: v1.0},
+  month        = dec,
+  year         = 2020,
+  publisher    = {Zenodo},
+  version      = {v1.0},
+  doi          = {10.5281/zenodo.4319936},
+  url          = {https://doi.org/10.5281/zenodo.4319936}
+}
+
+@software{Beck20c,
+  author       = {Marcus W Beck},
+  title        = {tbep-tech/nekton-dash: v1.0},
+  month        = jul,
+  year         = 2020,
+  publisher    = {Zenodo},
+  version      = {v1.0},
+  doi          = {10.5281/zenodo.3939236},
+  url          = {https://doi.org/10.5281/zenodo.3939236}
+}
+
+@software{Beck20d,
+  author       = {Marcus W Beck and
+                  Michael R. Wessel},
+  title        = {tbep-tech/tidalcreek-dash: v1.0},
+  month        = mar,
+  year         = 2020,
+  publisher    = {Zenodo},
+  version      = {v1.0},
+  doi          = {10.5281/zenodo.3701560},
+  url          = {https://doi.org/10.5281/zenodo.3701560}
+}
+
+@Manual{Chang21,
+    title = {shiny: Web Application Framework for {R}},
+    author = {Winston Chang and Joe Cheng and JJ Allaire and Carson Sievert and Barret Schloerke and Yihui Xie and Jeff Allen and Jonathan McPherson and Alan Dipert and Barbara Borges},
+    year = {2021},
+    note = {R package version 1.6.0},
+    url = {https://CRAN.R-project.org/package=shiny},
+  }
+
 @techreport{Karlen20,
 	author={{D.J. Karlen, T. Dix, B.K. Goetting, S.E. Markham, K.Campbell, J. Jernigan, J.Christian, K. Martinez, A. Chacour}},
 	year={2020},
@@ -62,5 +115,6 @@
   author={M. R. Wessel and J. R. Leverone and M. W. Beck and E. T. Sherwood and J. Hecker and S. West and A. Janicki},
   journal={Estuaries and Coasts},
   year={2021},
-  volume={In press}
+  volume={In press},
+  doi={10.1007/s12237-021-00974-7}
   }
