### 2021-08-18 1cd5c3b
widen scope of lit review in paper, fix #44
