+++ paper/paper.bib
@@ -67,6 +67,18 @@
     url = {https://CRAN.R-project.org/package=shiny},
   }
 
+@Manual{DeCicco21,
+    author = {Laura A. {De Cicco} and David Lorenz and Robert M. Hirsch and William Watkins and Mike Johnson},
+    title = {dataRetrieval: R packages for discovering and retrieving water data available from {U.S.} federal hydrologic web services},
+    publisher = {U.S. Geological Survey},
+    address = {Reston, VA},
+    version = {2.7.9},
+    institution = {U.S. Geological Survey},
+    year = {2021},
+    doi = {10.5066/P9X4L3GE},
+    url = {https://code.usgs.gov/water/dataRetrieval},
+  }
+
 @techreport{Karlen20,
 	author={{D.J. Karlen, T. Dix, B.K. Goetting, S.E. Markham, K.Campbell, J. Jernigan, J.Christian, K. Martinez, A. Chacour}},
 	year={2020},
@@ -109,6 +121,14 @@
 	misc={},
 	url={https://drive.google.com/file/d/1HgnP5dC2LQqffc6Bq6klEuSUcNC2WRiZ/view?usp=drivesdk},
 }
+
+@Manual{Thorley18,
+    title = {wqindex: Water Quality Index Calculation for {B}ritish {C}olumbia},
+    author = {Joe Thorley and Colin Millar and Andy Teucher},
+    year = {2018},
+    note = {R package version 0.3.1.9001},
+    url = {https://github.com/bcgov/wqindex},
+  }
 
 @article{Wessel21,
   title={Developing a water quality assessment framework for southwest Florida tidal creeks},
