+++ paper/paper.Rmd
@@ -54,7 +54,9 @@ options(tibble.width = 60, tibble.print_max = 5, tibble.print_min = 5)
 
 Many environmental programs report on the status and trends of natural resources to inform management decisions for protecting or restoring environmental condition.  The National Estuary Program (NEP) in the United States is one example of a resource management institution focused on "estuaries of national significance" that provides place-based solutions to managing coastal resources.  There are 28 NEPs in the United States, each with similar but location-specific programmatic goals to address environmental challenges related to water quality, alteration of hydrologic flows, invasive species, climate change,  declines in fish and wildlife populations, pathogens and other contaminants, and stormwater management.  A critical need of each NEP is the synthesis of data from disparate sources that can inform management response to address these environmental challenges. 
 
-The Tampa Bay Estuary Program (TBEP) in Florida, USA is responsible for developing and implementing a place-based plan to sustain historical and future progress in the restoration of Tampa Bay [@tbep1017].  The needs of TBEP for reporting on indicators of environmental condition are similar to other environmental organizations.  Multiple local and regional partners collect data that are used for different reporting products.  Without data synthesis tools that are transparent, accessible, and reproducible, NEP staff and colleagues waste time and resources compiling information by hand.  The `tbeptools` R software package can be used for routine development of reporting products, allowing for more efficient use of limited resources and a more effective approach to communicate research to environmental decision-makers.  The package imports required data directly from sources, removing the need to manually obtain information prior to reporting.  Functions in `tbeptools` also support the creation of content for interactive, online dashboards that can facilitate more informed decisions without requiring an intimate understanding of the R programming language or the methods for analysis.       
+The Tampa Bay Estuary Program (TBEP) in Florida, USA is responsible for developing and implementing a place-based plan to sustain historical and future progress in the restoration of Tampa Bay [@tbep1017].  The needs of TBEP for reporting on indicators of environmental condition are similar to other environmental organizations.  Multiple local and regional partners collect data that are used for different reporting products.  Without data synthesis tools that are transparent, accessible, and reproducible, NEP staff and colleagues waste time and resources compiling information by hand.  The `tbeptools` R software package can be used for routine development of reporting products, allowing for more efficient use of limited resources and a more effective approach to communicate research to environmental decision-makers. Functions in `tbeptools` also support the creation of content for interactive, online dashboards that can facilitate more informed decisions without requiring an intimate understanding of the R programming language or the methods for analysis.       
+
+The `tbeptools` package also addresses challenges associated with data retrieval and assessment of environmental data relative to important policy or management targets.  For each environmental indicator, functions are included to import required data directly from sources, removing the need to manually obtain information prior to reporting.  These functions are integrated into summary report cards that are generated automatically through continuous integration services (i.e., [GitHub Actions](https://github.com/features/actions)) that free the analyst from external downloads, analysis, and copying of results that can introduce errors in reporting.  Similar packages provide seamless access to database services [e.g., `dataRetrieval`, @DeCicco21], but few packages link these data sources directly to analysis and reporting as in `tbeptools` [but see `wqindex`, @Thorley18].  Management targets and regulatory thresholds based on summary assessmenta of data, either direct from sources or included as supplementary data in the package, are also hard-coded into the functions. 
 
 # Statement of need
 
