### 2021-08-27 tbep-tech/tbeptools#47 @richardsc
[JOSS Review] Consider more verbose function names
This comment is really about coding style, and while I hope the authors consider it seriously I don't consider this to be a blocking issue. I recognize that changing function names is a major breaking change, especially for a package that is as fully-developed as this one, and likely has a long history of use already within the institutions for which it was designed.

I like the consistency in the function verb prefixes (e.g. `read_`, `show_`, `anlz_`) however as an unfamiliar user I find it hard to parse the spectific intentions of the functions from the suffixes.

In general, I would prefer to see fully descriptive function names -- e.g. `analyze_` instead of `anlz_`, `water_quality` instead of `wq`, etc. As an example `anlz_avedat()` is completely opague to me as a function name -- the documentation says that it "Estimates annual means", so why not call it `analyze_annual_means()`. The argument against this is that it's a lot of typing, but in this day and age of good IDEs and code completion I prefer real words instead of confusing abbreviations.

Another thing to consider is that the mix of `snake_case` for the prefixes and *no* case convention for the suffixes is also hard to parse. I think that much of the functionality of the package would be more easily deduced if the functions were named as above, e.g. `analyze_annual_means()` or `analyze_site_attainments()`.

(Linking openjournals/joss-reviews#3485 )
