### 2021-08-31 #3485 @richardsc
Sorry for the delay in completing this. Summer, pandemic, etc etc ...

I had the benefit of reviewing after @paleolimbot , and my early review comments mirrored many of his. The paper is much improved by the changes since the first draft, and I think is acceptable for publication. I have two points for consideration, but these are not blocking in any way.

1. I made an issue on the repo (tbep-tech/tbeptools#47) to consider using more consistent and verbose function names. This is partly personal style preference, but there is so much functionalilty in the package (which is great!) that it is a bit hard to intuit what all the abbreviations mean.

2. Similar to @paleolimbot, I think a discussion of how this package could be used and/or generalized to be more broadly useful to the NEP community (or really *any* estuarine scientist) would improve the scope of the paper. I don't think that generalization has to be done here (it is a big undertaking), but would strengthen the potential utility of the package beyond users specifically interested in *only* the Tampa Bay Estuary. This point is touched on in the current version (including pointing out the `peptools` package), but I think it could be expounded upon a little further. Does every estuary need its own package to handle data import and analysis/reporting? The answer could be "yes", which is ok, but some thought about a broader approach would be nice.

All said, this is a nice paper and a great package. Well done @fawda123!
