### 2021-08-31 #3485 @fawda123
Thanks @richardsc, I really appreciate the comments.  I agree some of the function names could be more intuitive and I will work on some text revisions to discuss generalization/needs of other estuarine scientists.
