### 2021-08-31 #3485 @paleolimbot
I just did a read-through on the paper and revised README and like the changes. No further acceptance blockers from me, although I would put an `# Example` header before your example since it's not really in the `# Statement of need` section.
