### 2021-09-03 497d375
added sgseg data, minor doc update
