+++ R/sgseg.R
@@ -15,6 +15,7 @@
 #' \dontrun{
 #' library(sf)
 #' library(tidyverse)
+#' library(tools)
 #'
 #' prj <- 4326
 #'
@@ -25,7 +26,14 @@
 #'   ) %>%
 #'   st_transform(crs = prj) %>%
 #'   select(segment = SEAGRASSSE) %>%
-#'   mutate(segment = tolower(segment))
+#'   mutate(
+#'      segment = tolower(segment),
+#'      segment = case_when(
+#'         segment == 'terra ciea' ~ 'Terra Ceia',
+#'         T ~ segment
+#'      ),
+#'      segment = toTitleCase(segment)
+#'   )
 #'
 #' # save
 #' save(sgseg, file = 'data/sgseg.RData', compress = 'xz')
