+++ man/sgmanagement.Rd
@@ -1,0 +1,44 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/sgmanagement.R
+\docType{data}
+\name{sgmanagement}
+\alias{sgmanagement}
+\title{Seagrass management areas for Tampa Bay}
+\format{
+A simple features \code{\link[sf]{sf}} object (MULTIPOLYGON) with 30 features and 1 field, +proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs
+\describe{
+  \item{areas}{int}
+}
+}
+\usage{
+sgmanagement
+}
+\description{
+Seagrass management areas for Tampa Bay
+}
+\details{
+These polygons are seagrass management areas for Tampa Bay that provide a finer division of areas within major segments (\code{\link{tbseg}}) having relevance for locations of seagrass beds.
+}
+\examples{
+\dontrun{
+library(sf)
+library(tidyverse)
+library(tools)
+
+prj <- 4326
+
+# create sf object of boundaries
+sgmanagement <- st_read(
+  dsn = '~/Desktop/TBEP_SG_MA_FINAL_Project.shp',
+  drivers = 'ESRI Shapefile'
+  ) \%>\%
+  st_transform(crs = prj) \%>\%
+  select(areas = TBEP_SG_MA)
+
+# save
+save(sgmanagement, file = 'data/sgmanagement.RData', compress = 'xz')
+
+}
+}
+\concept{data}
+\keyword{datasets}
