### 2021-09-03 db0bbba
fix doc typos
