+++ R/tbseg.R
@@ -8,7 +8,7 @@
 #'   \item{bay_segment}{chr}
 #' }
 #'
-#' @description Note that these boundaries not used for formal analysis and only for a visual aids in mapping.
+#' @description Note that these boundaries are not used for formal analysis and are only used as visual aids in mapping.
 #'
 #' @concept data
 #'
