+++ man/tbseg.Rd
@@ -15,7 +15,7 @@ A simple features \code{\link[sf]{sf}} object (POLYGON) with 4 features and 2 fi
 tbseg
 }
 \description{
-Note that these boundaries not used for formal analysis and only for a visual aids in mapping.
+Note that these boundaries are not used for formal analysis and are only used as visual aids in mapping.
 }
 \details{
 Spatial data object of Tampa Bay segments
