+++ R/sgseg.R
@@ -29,7 +29,7 @@
 #'   mutate(
 #'      segment = tolower(segment),
 #'      segment = case_when(
-#'         segment == 'terra ciea' ~ 'Terra Ceia',
+#'         segment == 'terra ciea bay' ~ 'Terra Ceia Bay',
 #'         T ~ segment
 #'      ),
 #'      segment = toTitleCase(segment)
