### 2021-09-08 595af43
drop z attribute from sgmanagement
