+++ R/sgmanagement.R
@@ -25,7 +25,8 @@
 #'   drivers = 'ESRI Shapefile'
 #'   ) %>%
 #'   st_transform(crs = prj) %>%
-#'   select(areas = TBEP_SG_MA)
+#'   select(areas = TBEP_SG_MA) %>%
+#'   st_zm()
 #'
 #' # save
 #' save(sgmanagement, file = 'data/sgmanagement.RData', compress = 'xz')
