### 2021-09-08 f82012f
fixing proj for sgseg and sgmanagement to epsg 2882
