+++ paper/paper.md
@@ -58,6 +58,14 @@ The `tbeptools` package also addresses challenges associated with data retrieval
 
 The `tbeptools` R package was developed to automate data synthesis and analysis for many of the environmental indicators for Tampa Bay, with more general application to commonly available datasets for estuaries.  The functions in the package were developed to extract methods from existing technical documents and to make them available in an open source programming environment.  By making these tools available as an R package, routine assessments are now accomplished more quickly and other researchers can use the tools to develop more specific analysis pipelines.  
 
+Most of the NEPs do not have analysis software to operationalize data import, analysis, and plotting for reporting.  Recently, a similar software package, `peptools` [@Beck21a], was developed for the Peconic Estuary Partnership (New York, USA) using many of the functions in `tbeptools` to develop reporting products for a new water quality monitoring program.  This successful technology transfer demonstrates the added value of presenting these methods in an open source environment available for discovery and reuse by others.  We expect other NEPs to begin using these tools as their application becomes more widespread among estuarine researchers.
+
+Beyond the NEPs, `tbeptools` is an effective example of an R package for implementing technical methods in existing literature and reports that can be used to support environmental monitoring and assessment needs for science-based decisions.  To this end, the `tbeptools` package was also created to support the development of online dashboards created in R Shiny [@Chang21].  Dashboards are powerful tools to increase accessibility for end users to engage with scientific products without the need to understand technical details in their creation.  However, providing the underlying methods as source code in an R package increases transparency and reproducibility of reporting products if users require a more detailed understanding of how the content was created.  Currently, the `tbeptools` package supports dashboards created by TBEP for the assessment of [water quality](https://shiny.tbeptech.org/wq-dash/) [Figure \ref{fig:dashex}, @Beck20a], [seagrasses](https://shiny.tbep.org/seagrasstransect-dash/) [@Beck20b], [nekton communities](http://shiny.tbeptech.org/nekton-dash) [@Beck20c], and [tidal creeks](https://shiny.tbep.org/tidalcreek-dash/) [@Beck20d].  Resource management agencies or similar institutions could follow this approach to facilitate development of front-end products for more informed decision-making. 
+
+![The TBEP water quality dashboard, demonstrating use of the `tbeptools` R package to generate summary plots for specific bay segments.\label{fig:dashex}](paper_files/figure-latex/dashex.JPG){width='100%'}
+
+# Example usage
+
 The function names were chosen with a typical analysis workflow in mind, where functions are available to `read` data from a source (typically from an online repository or stable URL), `anlz` to analyze the imported data using methods in existing technical documents or published papers, and to `show` the results as a summary graphic for use by environmental managers.  The functions are used to report on water quality [@Beck21b], fisheries [@Schrandt21], benthic condition [@Karlen20], tidal creeks [@Wessel21], and seagrass transect data [@Sherwood17]. The [vignettes](https://tbep-tech.github.io/tbeptools/articles/intro.html) for the package are topically organized to describe the functions that apply to each of the indicators.
 
 The following example demonstrates use of a subset of the functions for water quality data to read a file from the Hillsborough County Environmental Protection Commission long-term monitoring dataset (available from <https://www.tampabay.wateratlas.usf.edu/>), analyze monthly and annual averages by major bay segments of Tampa Bay, and plot an annual time series for one of the bay segments.
@@ -128,14 +136,8 @@ show_thrplot(wqdat, bay_segment = 'HB', yrrng = c(1975, 2020))
 ![](paper_files/figure-latex/thrplotex-1.jpeg)<!-- --> 
 
 
[…]
