### 2021-09-10 #3485 @fawda123
@paleolimbot I just added an `# Example usage` section https://github.com/tbep-tech/tbeptools/commit/ec595837d30d278c2fd1cb2a338eb76f34594f94
