### 2021-09-10 3514aa9
suffix description to readme and paper #47
