### 2021-09-10 8a9dc63
change example usage section in paper to usage
