### 2021-09-10 tbep-tech/tbeptools#47 @fawda123
Thanks @richardsc.  I fully agree that function names should be as intuitive as possible.  Since hindsight is of course 20/20, I realize I could have done a better job for the function names in tbeptools.  I did consider whether changing the names was practical, but as it stands, this package is used across several dozen repositories on our group web page (https://github.com/tbep-tech) and in CI actions through GitHub Actions.

Rather than risk breaking lots of time-sensitive things (some of which are under regulatory compliance, e.g., https://tbep-tech.github.io/tbnmc-compliance-assessment-2020/), I figured a good compromise was to provide some descriptive text on the web page and to point the reader there in the manuscript.  I hope that's a good compromise.  Again, I appreciate your comments on the package and paper.  I'm sure we've got a better product as a result of these review!
