### 2021-09-10 #3485 @fawda123
@kthyng @richardsc @paleolimbot I'm done addressing the comments from all.  I really appreciate the thoughtful comments on this package and paper.  I think we've got a much better product as a result.

For the first comment from @richardsc, I've added some descriptive text describing the suffixes (see my response here https://github.com/tbep-tech/tbeptools/issues/47).  For the second comment (i.e., "A discussion how the package can be used beyond the NEPs..."), I feel that the revisions in response @paleolimbot address that concern (see the commit here https://github.com/tbep-tech/tbeptools/commit/227915bc7024de827f3f2f83bf482db7a82755d1).  I'm happy to revise further, if needed, but I am done for now.  All, please let me know if these changes are sufficient.
