### 2021-09-13 d5546a8
fix to sgmanagement for area 6 boundary
