### 2021-09-16 #3485 @kthyng
@fawda123 Can you verify what version of the software should be associated with your JOSS publication? Also, can you make an archive of the software at a place like Zenodo and then report the doi here to associate with your paper? Please modify the metadata of the archive as needed so that the title and author list exactly match your JOSS paper.
